open_sdk/streamlit_app/requirements.txt
# Core Streamlit
streamlit>=1.28.0

# Langchain & AI
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-core>=0.1.0
google-generativeai>=0.3.0

# OpenAI (for Gemini compatibility)
openai>=1.50.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
folium==0.14.0

# Geospatial
geopandas>=0.13.0
shapely>=2.0.0

# Image Processing
Pillow>=10.0.0

# HTTP & Requests
requests>=2.31.0
httpx>=0.24.0

# Geocoding
geopy>=2.3.0

# Configuration & Environment
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Async & Threading
asyncio-mqtt>=0.13.0

# File I/O
aiofiles>=23.0.0

# Monitoring (Optional)
langfuse>=2.50.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
python-multipart>=0.0.6
tiktoken>=0.5.0
